fix(agents): keep core green without the chat wheel + fix stale hub tests
CI's framework-only jobs (Unit Tests on `.[api]`, Security Tests on `.[dev,rag]`)
run without gaia-agent-chat, which surfaced regressions the wheel-present local
run masked:

- The dynamic-tools UI toggle 500'd without the wheel: `dynamic_tools_env_override`
  had moved into the chat wheel but the core UI settings router imports it. Moved
  the helper to the core `tool_loader` (its feature home) and re-export it from
  the wheel's agent.py — the router no longer depends on the wheel, one truthy
  set preserved.
- `test_chat_dynamic_tools` stubbed faiss/sentence_transformers as MagicMocks
  BEFORE the `importorskip`, so a skip (no wheel) leaked the stubs into the run
  and broke later memory/code-index tests. Moved the importorskip above the stub
  block so a skip leaves sys.modules untouched.
- Guarded `test_dynamic_tool_filtering` (transitively builds a ChatAgent via
  `build_doc_agent_skeleton`) with importorskip.
- Updated stale hub tests for the migration: chat/doc are no longer reserved ids
  or builtins, so `test_reserved_id_rejected` keeps only `builder`,
  `test_uninstall_builtin_refused` targets `builder`, and the agent-hub-metadata
  tests run full discovery + importorskip for the now-installed chat/doc agents.
- test_security.yml installs the chat wheel (the verify_*.py guardrail scripts
  instantiate ChatAgent).
