+++ tests/unit/test_dynamic_tool_filtering.py
@@ -18,7 +18,13 @@ from __future__ import annotations
 
 import json
 
-from gaia.eval.tool_cost import build_doc_agent_skeleton
+import pytest
+
+# build_doc_agent_skeleton builds a doc-profile ChatAgent, which ships in the
+# standalone gaia-agent-chat wheel (#1102); skip when a framework-only env lacks it.
+pytest.importorskip("gaia_agent_chat")
+
+from gaia.eval.tool_cost import build_doc_agent_skeleton  # noqa: E402
 
 
 def _agent():
