+++ tests/unit/test_hub_manifest.py
@@ -281,7 +281,9 @@ def test_valid_id_accepted(good_id):
     assert AgentManifest.from_dict(data).id == good_id
 
 
-@pytest.mark.parametrize("reserved", ["chat", "builder", "gaia-lite", "doc"])
+# Only ``builder`` remains a framework builtin; chat/doc/data/web/email migrated
+# to standalone hub wheels (#1102) and are no longer reserved ids.
+@pytest.mark.parametrize("reserved", ["builder"])
 def test_reserved_id_rejected(reserved):
     data = dict(VALID_PYTHON_MANIFEST, id=reserved)
     with pytest.raises(ManifestError, match="reserved"):
